move asset badge and mappings for assetType and marketState to separate file
